Refactor calendar
